pyproject.toml
[tool.poetry]
name = "github_notifications_bot"
version = "0.1.0"
description = ""
authors = ["vlzh <vlzh@pm.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
PyGithub = "^1.51"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

